Add HM Land Registry INSPIRE parcel layer to UK Context Builder
UK Context Builder parcels layer now queries the free HM Land Registry
INSPIRE WFS (inspire.landregistry.gov.uk) for registered freehold property
boundaries in England and Wales, falling back to OpenStreetMap cadastral
data if the service is unavailable. The dry run / preview report shows
which source was used.

Also includes general script and UI updates across the extension.

+++ WWPTools.tab/2. Context.panel/MassIDTool.pushbutton/masstofloor-script.py
@@ -1,3 +1,4 @@
+from System import Int64
 import clr
 import traceback
 
@@ -73,7 +74,7 @@ def _set_param_value(param, value):
                 param.Set(value)
                 return True
             if isinstance(value, int):
-                param.Set(ElementId(value))
+                param.Set(ElementId(Int64(value)))
                 return True
     except Exception:
         return False
